### 2025-08-18 #720 @maelle
babelquarto: render a Quarto multilingual project, book or website
Submitting Author Name: Maëlle Salmon
Submitting Author Github Handle: @maelle
Repository:  https://github.com/ropensci-review-tools/babelquarto/
Version submitted: 0.0.1
Submission type: Standard

Version accepted: TBD

---

```
Package: babelquarto
Title: Renders a Multilingual Quarto Book
Version: 0.0.1
    person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = c("cre", "aut"),
    person("Johannes", "Ranke", role = "ctb"),
    person("Joel H.", "Nitta", , "joelnitta@gmail.com", role = "ctb",
    person("Pascal", "Burkhard", , "pascal.burkhard@gmail.com", role = "aut",
    person("rOpenSci", role = "fnd",
           comment = c(ROR = "019jywm96"))
  )
Description: Automate rendering and cross-linking of Quarto books following a
    prescribed structure.
Roxygen: list(markdown = TRUE, roclets = c("collate", "rd", "namespace", "devtag::dev_roclet"))
RoxygenNote: 7.3.2.9000
    brio,
    cli,
    fs,
    quarto,
    sys,
    whoami,
    withr,
    xml2,
    yaml
    servr,
    quarto
```

	- [x] rOpenSci tools

It was developed by rOpenSci staff.
Based on https://devdevguide.netlify.app/softwarereview_policies#aims-and-scope, it also falls under the "traduction" category as it allows to publish multilingual documentation.

This package is useful for Quarto users who need to publish a multilingual Quarto website or book, as it is [not possible in native Quarto yet](github.com/quarto-dev/quarto-cli/issues/275).
The package can be used to publish any kind of Quarto website/book but [use cases](https://github.com/search?q=babelquarto%3A+path%3A**%2F_quarto.yml+NOT+repo%3Aropensci%2Fdev_guide+NOT+repo%3Aropensci-review-tools%2Ftranslation_guide%2F&type=code&ref=advsearch&p=2) include, beyond several rOpenSci guides, [Joel Nitta's personal website](https://joelnitta.com/), a [R handbook for epidemiologists](https://www.epirhandbook.com/en/), a website for [training courses in epidemiology](https://epicentre-msf.github.io/repicentre/).

No, not to my knowledge. 😄

N/A
