### 2025-08-18 d4e3926
fix: make sitemap fixing conditional on there being a sitemap
Fix #108
