+++ NEWS.md
@@ -1,3 +1,4 @@
 # babelquarto 0.0.1
 
 * Initial rOpenSci submission.
+* Fix handling on sitemap: it needed to be conditional on there being a sitemap.
