### 2025-08-18 17852ba
ci: re-add steps removed in #102
what a mystery
