### 2025-08-18 3faabca
clarifications on site url for render
