### 2025-09-11 51650a4
docs: add rOpenSci software review badge
