### 2025-10-05 #720 @joaojcorreia
---

Thank you for the opportunity to review **babelquarto**. As someone who frequently works with multilingual projects, I find this package incredibly useful for simplifying the repetitive tasks and streamlining workflows of working with multiple languages.

I believe the **statement of need** could be more assertive and clearer, better highlighting the package's value and potential impact for multi-language projects.

During my review, I encountered a few issues, which I detail below. While these are not insurmountable for experienced users, they could pose challenges for those of us with less experience.

#### Suggestions for Improvement

- **Feature Request:**
  It would be highly beneficial to allow users to define the formatting of references. Different academic fields and industries adhere to distinct formatting standards, so the ability to customize reference formatting would greatly enhance the package's flexibility and appeal.

- **Documentation Improvement:**
Further guidance should be provided in the vignettes on how to handle the `site-url` configuration in the YML file. Inexperienced users may be surprised to find that the site does not work as expected after rendering and navigating to the second language. Ideally, the vignette should include an example of how to customize the `site-url` for local use, as well as instructions for setting it up for online publishing.

#### Issues

- **Adding additional languages:**
  It was possible to add a second language using the `register_further_languages` function. However, when attempting to add a third language, or rendering the project, the following error occurs:

  ```error
  register_further_languages("it", project_path = "book")
      Error in yaml.load(string, error.label = error.label, ...) :
      (book/_quarto.yml) Scanner error: mapping values are not allowed in this context at line 39, column 12
  ```

  **Root cause:**
  The error is caused by the function not removing the previous `languages: []` line in the yml file. This creates a conflict when adding another language and during rendering, resulting in the above error.

- **Rendering error:**
Occasionally, the following error occurs when rendering the book/site:

```error
babelquarto::render_book(file.path(parent_dir, project_dir))
Error: [EPERM] Failed to remove 'mypath/test_babelquarto/book/_book/site_libs/quarto-search': operation not permitted
```

**Observation:**
The pattern for this error is not yet clear, but when it happens, manually deleting the rendered book files resolves the issue, and the book renders normally afterward.
