### 2025-10-12 9a52d3b
Correct vignette index entry
