+++ vignettes/configuration.qmd
@@ -41,7 +41,7 @@ By default, {babelquarto} will have the following options under the `babelquarto
 | `mainlanguage` | The main language of the project. |
 | `languages` | An array of the additional languages (not including the main language). |
 
-In addition to the `babelquarto` key, you will see language specific keys for `title`, `description` and `author`. These keys allow you to have a specific title, description or author in for each additional language.
+In addition to the `babelquarto` key, you will see language specific keys for `title`, `description` and `author`. These keys allow you to have a specific title, description or author for each additional language.
 
 ### Customizing the language menu labels
 
