### 2025-10-13 b441a39
A few tiny fixes (#110)
* Delete unnecessary (and non-standard) pkdown.yml

* Correct vignette index entry

* Remove unnecessary `</div>`

* Fix type

* Consistent case
