+++ vignettes/custom-templates.qmd
@@ -3,7 +3,7 @@ title: "Custom templates"
 description: >
   How to have multilingual parts in your custom templates
 vignette: >
-  %\VignetteIndexEntry{Convert an existing project}
+  %\VignetteIndexEntry{Custom templates}
   %\VignetteEngine{quarto::html}
   %\VignetteEncoding{UTF-8}
 ---
@@ -41,5 +41,4 @@ $endif$
 $if(lang-fr)$
 <meta name="publisher" content="Publisher for the French part">
 $endif$
-</div>
 ```
