### 2025-10-21 f2b6024
docs: specify `render_book()`'s default value for `project_path`
