### 2025-10-21 2feafec
docs: clarify the package's value
