### 2025-10-21 552e78c
fix: make it possible to `register_further_languages()` for more languages in a second step
