### 2025-10-21 ebb53da
docs: add how to preview earlier in the vignette.
