### 2025-10-21 dc8a3a7
fix: make it possible to `register_further_languages()` for more lang… (#111)
* fix: make it possible to `register_further_languages()` for more languages in a second step

* news
