+++ NEWS.md
@@ -1,3 +1,14 @@
+# babelquarto (development version)
+
+## Documentation improvements
+
+- clarify the default value of `project_path` in `render_book()`, in the "Get started" vignette.
+- strengthen the case for the package in the README.
+
+## Bug fixes
+
+- make it possible to `register_further_languages()` for more languages in a second step.
+
 # babelquarto 0.0.1
 
 * Initial rOpenSci submission.
