+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: Automate rendering and cross-linking of Quarto books following a
 License: MIT + file LICENSE
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE, roclets = c("collate", "rd", "namespace", "devtag::dev_roclet"))
-RoxygenNote: 7.3.2.9000
+RoxygenNote: 7.3.3.9000
 Imports: 
     brio,
     cli,
