### 2025-10-21 96b3cbc
docs: mention `site-url` param of `quarto_multilingual_` functions
