### 2025-10-21 528e438
docs: clarify what publishing means in practice
