### 2025-10-21 f48d735
docs: add default for `languagelinks` in table
