### 2025-10-21 a4ff674
docs: clarify `languagecodes`
