### 2025-10-21 6420102
docs: clarify `site-url` in configuration vignette.
