### 2025-10-21 dae5775
docs: document how to get no icon, in configuration vignette.
