+++ NEWS.md
@@ -13,6 +13,7 @@
 - new section in the Get started vignette, "publishing your project".
 - clarify `languagecodes` in configuration vignette.
 - clarify `site-url` in configuration vignette.
+- document how to get no icon, in configuration vignette.
 
 ## Bug fixes
 
