### 2025-10-21 4efe76d
docs: clarify profile naming convention, in configuration vignette.
