+++ NEWS.md
@@ -14,6 +14,7 @@
 - clarify `languagecodes` in configuration vignette.
 - clarify `site-url` in configuration vignette.
 - document how to get no icon, in configuration vignette.
+- clarify profile naming convention, in configuration vignette.
 
 ## Bug fixes
 
