### 2025-10-21 b1bdc64
fix: avoid unnecessary empty lines in Quarto config when using `register_main_language()`.
