### 2025-10-21 e83c125
feat: Add verbosity to `register_` functions, that one can turn off through an option
