+++ tests/testthat/setup.R
@@ -1,0 +1,1 @@
+withr::local_options(babelquarto.quiet = TRUE, .local_envir = teardown_env())
