+++ tests/testthat/_snaps/ui.md
@@ -1,0 +1,14 @@
+# cli_alert_success() works
+
+    Code
+      cli_alert_success("hihihi")
+    Message
+      v hihihi
+
+# cli_alert_info() works
+
+    Code
+      cli_alert_info("hihihi")
+    Message
+      i hihihi
+
