+++ tests/testthat/_snaps/register/_quarto.yml
@@ -27,3 +27,4 @@ babelquarto:
     text: "Version in en"
   mainlanguage: 'en'
 lang: en
+
