+++ tests/testthat/_snaps/register/_quarto2.yml
@@ -43,3 +43,4 @@ description-fr: description in fr
 author-es: author in es
 author-fr: author in fr
 lang: en
+
