### 2025-10-21 7aead89
docs: clarify the default value of `project_path` in `register_` functions, in the "Converting an existing project" vignette.
