### 2025-10-21 e6350a9
docs: clarify `BABELQUARTO_CI_URL`, in render-with-ci vignette.
