### 2025-10-21 1b416c0
docs: clarify config file
