+++ vignettes/render-with-ci.qmd
@@ -14,7 +14,7 @@ You can find more information about publishing with CI in the [Quarto documentat
 
 ### Site URL
 
-When you render a project in a CI context, you can set the URL of the publishing website in an environment variable (if different from the one in the configuration file,
+When you render a project in a CI context, you can set the URL of the publishing website in an environment variable (if different from the one in the `_quarto.yml` configuration file,
 as a preview for instance).
 You can do that with the `BABELQUARTO_CI_URL` environment variable.
 
