### 2025-10-21 efff962
docs: add publishing section here too
