### 2025-10-21 9328ba6
docs: improve return value
