### 2025-10-21 eafd563
docs: remove stray character
