### 2025-10-21 fe1f36f
docs: rm useless example
