+++ R/render.R
@@ -35,11 +35,6 @@
 #' directory <- withr::local_tempdir()
 #' quarto_multilingual_book(parent_dir = directory, project_dir = "blop")
 #' render_book(file.path(directory, "blop"))
-#' \dontrun{
-#' if (require("servr") && rlang::is_interactive()) {
-#'   servr::httw(file.path(directory, "blop", "_book"))
-#' }
-#' }
 #'
 #' @rdname render
 render_book <- function(
