+++ DESCRIPTION
@@ -10,7 +10,9 @@ Authors@R: c(
     person("Pascal", "Burkhard", , "pascal.burkhard@gmail.com", role = "aut",
            comment = c(ORCID = "0009-0001-5504-5084")),
     person("rOpenSci", role = "fnd",
-           comment = c(ROR = "019jywm96"))
+           comment = c(ROR = "019jywm96")),
+    person("Ella", "Kaye", role = "rev",
+           comment = c("Ella reviewed the package (v. 0.0.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/720>.", ORCID = "0000-0002-7300-3718"))
   )
 Description: Automate rendering and cross-linking of Quarto books following a 
     prescribed structure.
