### 2025-10-21 #720 @maelle
Thanks @EllaKaye, enjoy your leave! You're now [listed in DESCRIPTION](https://github.com/ropensci-review-tools/babelquarto/commit/3e5f2121a8960483ceca1178525a7bd878eb4846).
