### 2025-10-21 6cac01b
docs: add @joaojcorreia as "rev"
