### 2025-10-21 #720 @maelle
Thanks @joaojcorreia, you are now [listed in DESCRIPTION](https://github.com/ropensci-review-tools/babelquarto/commit/6cac01bc9dcd8a6dd7335558e086b822678726f9).

> I tried to manually change the format, and couldn't. Might have been me that made mistake. However, in some contexts, it might be useful to be able to setup a format in one language and another in a different language. I understand that this might be a niche use-case and not worth the time.

Would you be able to create a small reprex? To make things different in one language, maybe you could use a profile. But what do you mean by citation format? The CSL file? https://quarto.org/docs/authoring/citations.html#bibliography-files
