### 2025-10-26 #720 @ronnyhdez
Hi @EllaKaye and @joaojcorreia!

Thanks for the reviews and follow-up comments!

Whenever you’re ready, please let us know if you’d require any further changes. Otherwise, if you’re satisfied with the updates in the package, please indicate your approval using [this template](https://devguide.ropensci.org/approval2template.html)
