### 2025-11-01 #720 @joaojcorreia
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 7
