### 2025-11-04 #720 @EllaKaye
@ronnyhdez and @maelle I've been working through Maëlle's changes and lovely to see some big and helpful changes, such as the interactive preview - a really excellent improvement. I do have a few more suggestions/requests but I'm afraid I haven't been able to finish going through everything today, and I'd rather keep all of those in one comment. Apologies for the delay. My schedule is rammed at the moment, and as much as I wish I can commit to finishing this in the next day or two, more realistically it's likely to be towards the end of next week. I hope that's OK.
