### 2025-11-16 #720 @ronnyhdez
Hi @EllaKaye,

Thanks so much for the update and for all the careful work you’re putting into the review! it’s greatly appreciated!

Just checking: will the upcoming “few more suggestions/requests” you mentioned be a continuation of the earlier feedback, or new points we haven’t seen yet?

Thanks again for keeping us updated and for all your time on this!
