### 2025-11-18 9a0ef26
fix: remove stray word
