+++ R/register.R
@@ -36,17 +36,7 @@ register_main_language <- function(main_language = "en", project_path = ".") {
   config_lines <- brio::read_lines(config_path)
 
   # avoid more than one empty lines at the end
-  if (!nzchar(utils::tail(config_lines, n = 1))) {
-    reps <- rle(config_lines)[["lengths"]]
-    how_many_empty <- reps[length(reps)]
-    if (how_many_empty > 1) {
-      config_lines <- config_lines[
-        1:(length(config_lines) - how_many_empty + 1)
-      ]
-    }
-  } else {
-    config_lines <- c(config_lines, "")
-  }
+  config_lines <- trim_end(config_lines)
 
   mainlang_config <- c(
     "  languagecodes:",
@@ -157,6 +147,7 @@ register_further_languages <- function(further_languages, project_path = ".") {
     after = which_codes
   )
 
+  config_lines <- trim_end(config_lines)
   brio::write_lines(config_lines, path = config_path)
 
   cli_alert_success(
