+++ vignettes/babelquarto.qmd
@@ -31,6 +31,8 @@ pak::pak("ropensci-review-tools/babelquarto")
 To start a multilingual book, use `quarto_multilingual_book()` with the `parent_dir` argument to specify where you want to create the project and the `project_dir` argument to specify the name of the project.
 The argument `main_language` is used to specify the main language of the project and `further_languages` lists all additional languages.
 
+You can set the `site-url` field to your project's URL directly by using the `site_url` parameter of `quarto_multilingual_website()`/`quarto_multilingual_book()`.
+
 ```{r}
 #| label: setup_book
 #| results: 'hide'
@@ -93,6 +95,8 @@ See the [Previewing your multilingual project section](#preview) for further det
 To start a multilingual website, use `quarto_multilingual_website()` with the `parent_dir` argument to specify where you want to create the project and the `project_dir` argument to specify the name of the project.
 The argument `main_language` is used to specify the main language of the project and `further_languages` lists all additional languages.
 
+You can set the `site-url` field to your project's URL directly by using the `site_url` parameter of `quarto_multilingual_website()`/`quarto_multilingual_book()`.
+
 ```{r}
 #| label: setup_website
 #| results: 'hide'
@@ -144,7 +148,6 @@ babelquarto::render_website(file.path(parent_dir, project_dir))
 We end up with a multilingual website. [Example](https://maelle.github.io/babelsite), [source](https://github.com/maelle/babelsite)
 
 Note that the `_quarto.yml` file created in this process has an example entry set for `site-url:`. Depending on how your multilingual project is being deployed, you may need to either remove or update this entry or alternatively, set the url during rendering by using the `site_url` argument in the render functions (e.g., `babelquarto::render_website(site_url = "https://your.blog.net/your-multilingual-book/")`).
-You can set the `site-url` field to your project's URL directly by using the `site_url` parameter of `quarto_multilingual_website()`/`quarto_multilingual_book()`.
 
 ## Previewing your multilingual project {#preview}
 
