### 2025-11-18 dd4f680
docs: add more about deploy
