+++ vignettes/render-with-ci.qmd
@@ -46,7 +46,7 @@ You need to publish the entire output folder, including the language folders.
 
 ## Examples
 
-Here are two examples of how you could render a multilingual book with CI using GitHub Actions:
+Here are a few examples of how you could render a multilingual book with CI using GitHub Actions:
 
 * A demo [Babelbook](https://nenuial.github.io/babelbook/) with its [workflow file](https://github.com/Nenuial/babelbook/blob/main/.github/workflows/publish.yml)
 * The [rOpenSci Packages: Development, Maintenance, and Peer Review](https://devguide.ropensci.org/index.html) book with its [workflow file](https://github.com/ropensci/dev_guide/blob/main/.github/workflows/scheduled-manual-main.yml)
