### 2025-11-18 3930dde
docs: add docs URL
