+++ DESCRIPTION
@@ -39,7 +39,7 @@ Depends:
 Suggests: 
     servr,
     testthat (>= 3.0.0)
-URL: https://github.com/ropensci-review-tools/babelquarto
+URL: https://docs.ropensci.org/babelquarto/, https://github.com/ropensci-review-tools/babelquarto
 BugReports: https://github.com/ropensci-review-tools/babelquarto/issues
 Config/testthat/edition: 3
 VignetteBuilder:
