### 2025-11-18 db2cb0d
chore: namespace `tail()`
