### 2025-11-19 #720 @ropensci-review-bot
Approved! Thanks @maelle for submitting and @EllaKaye, @joaojcorreia for your reviews! :grin:
